+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './modules/http/constants';
 export * from './modules/http/services';
 export * from './modules/http/interceptors';
 export * from './modules/http/errors';
+export * from './modules/http/types';